Reporte integral con datos de basico

+++ src/main/java/com/zonainmueble/reports/maps/here/isoline/Range.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class Range {
   private String type;
-  private Double value;
+  private Integer value;
 }
